Fixed double declaration
Fixed the InfoBase class issue with #ifndef
Also, creating a RuleBase object calls the InfoBase AND the RuleBase
constructor? Is that how it's supposed to work?

Also, use "make simple" to compile for now, not ideal but I can't get
the rest of the makefile working (use just "make" to see)